chore: remove job-cron-caller and registry-cron-caller directories, migrating to sync-function for API synchronization
