--- job-cron-caller/project.yml
@@ -1,27 +1,0 @@
-packages:
-  - name: job-sync
-    shared: false
-    environment: {}
-    parameters: {}
-    annotations: {}
-    functions:
-      - name: main
-        binary: false
-        main: "main"
-        runtime: nodejs:18
-        web: true
-        webSecure: false
-        parameters: {}
-        environment:
-          ADMIN_KEY: ${ADMIN_KEY}
-          SOKOSUMI_API_URL: ${SOKOSUMI_API_URL}
-        annotations: {}
-        limits: {}
-        triggers:
-          - name: trigger-job-sync
-            sourceType: scheduler
-            sourceDetails:
-              # every minute
-              cron: "* * * * *"
-              withBody:
-                name: "job-sync"
