--- job-cron-caller/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es2020",
-    "module": "es2020",
-    "moduleResolution": "node",
-    "outDir": "./packages/job-sync/main",
-    "rootDir": "./src",
-    "strict": true,
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true
-  },
-  "include": ["src/**/*", "eslint.config.mjs"],
-  "exclude": ["node_modules", "packages/**/*"]
-}
