--- registry-cron-caller/.gitignore
@@ -1,50 +1,0 @@
-# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
-
-# dependencies
-/node_modules
-/.pnp
-.pnp.*
-.yarn/*
-!.yarn/patches
-!.yarn/plugins
-!.yarn/releases
-!.yarn/versions
-
-# testing
-/coverage
-
-# next.js
-/.next/
-/dist/*
-
-# production
-/build
-
-# misc
-.DS_Store
-*.pem
-
-# debug
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-.pnpm-debug.log*
-
-# env files (can opt-in for committing if needed)
-.env*
-!.env.example
-
-
-# typescript
-*.tsbuildinfo
-next-env.d.ts
-
-.nimbella
-.deployed
-__deployer__.zip
-__pycache__/
-node_modules
-package-lock.json
-.DS_Store
-
-packages/registry-sync/main/*
