--- registry-cron-caller/.vscode/extensions.json
@@ -1,10 +1,0 @@
-{
-  "recommendations": [
-    "bradlc.vscode-tailwindcss",
-    "esbenp.prettier-vscode",
-    "dbaeumer.vscode-eslint",
-    "streetsidesoftware.code-spell-checker",
-    "prisma.prisma",
-    "lokalise.i18n-ally"
-  ]
-}
