--- registry-cron-caller/pnpm-workspace.yaml
@@ -1,3 +1,0 @@
-publicHoistPattern:
-  - "*eslint*"
-  - "*prettier*"
