--- registry-cron-caller/prettier.config.mjs
@@ -1,10 +1,0 @@
-/** @type {import('prettier').Config} */
-const config = {
-  trailingComma: "all",
-  tabWidth: 2,
-  semi: true,
-  singleQuote: false,
-  plugins: [],
-};
-
-export default config;
