+++ sync-function/.vscode/extensions.json
