+++ sync-function/pnpm-workspace.yaml
