+++ sync-function/prettier.config.mjs
